# app/docs/openapi.yaml
openapi: 3.0.0
servers:
  - description: Live API
    url: 'https://be.bestone.ws'
info:
  title: API DOCS
  license:
    name: API DOCS
  description: |
    **Steps for Basic Auth:**   
    1) first of all you need to have our  Dashboard username and password after that on this URL https://www.blitter.se/utils/basic-authentication-header-generator/ you can generate your basic Auth from it. 
    2) Now you have to add Authorization headers.

    **Timezone:**
    1) Get the Timezone from this URL https://timezonedb.com/time-zones
...
security:
  - BasicAuth: []
paths:
  /api/create/order:
    post:
      tags:
        - Last Mile
      summary: Create Order
      description: create order
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: null
                  httpStatus:
                    type: string
                    example: OK
                  data:
                    $ref: '#/paths/~1api~1create~1order/post/responses/200/content/application~1json/schema/data'
                data:
                  type: object
                  properties:
                    total:
                      type: integer
                      example: 1
                    totalCreated:
                      type: integer
                      example: 1
                    totalFailed:
                      type: integer
                      example: 0
                    orders:
                      type: array
                      items:
                        $ref: '#/paths/~1api~1create~1order/post/responses/200/content/application~1json/schema/orderrs'
                orderrs:
                  type: object
                  required:
                    - orders
                    - orderNumber
                    - orderTrackingNumber
                    - status
                    - msg
                    - order
                    - type
                  properties:
                    orderNumber:
                      type: string
                      example: 97260383
                    orderTrackingNumber:
                      type: string
                      example: API0000036012
                    status:
                      type: string
                      example: success
                    msg:
                      type: string
                      example: Order Created
                    order:
                      type: string
                      example: null
                    type:
                      type: string
                      example: DROP_SHIP
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: fail
                  message:
                    type: string
                    description: |
                      **Order Creation Failed Reasons:**
                      1) Customer city not found with code/name riyadh
                      2) Duplicate Reference no for client
                      3) Maximum Allowed Pieces for Client 100
                      4) Invalid customer name
                      5) Invalid customer customer-Addresses areaName
                      6) sender length should be between 1-50
                      7) weight of order must be greater than zero
                      8) Invalid customer mobile1
                    example: Customer city not found with code/name riyadh
                  httpStatus:
                    type: string
                    example: BAD_REQUEST
                  data:
                    type: string
                    example: null
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                orders:
                  $ref: '#/paths/~1api~1create~1order/post/requestBody/content/application~1json/schema/orders'
              orders:
                type: array
                items:
                  $ref: '#/paths/~1api~1create~1order/post/requestBody/content/application~1json/schema/by'
              by:
                type: object
                required:
                  - referenceNumber
                  - pieces
                  - description
                  - codAmount
                  - paymentType
                  - collectTax
                  - clientComments
                  - value
                  - sender
                  - weight
                properties:
                  referenceNumber:
                    type: string
                    example: BS-01
                  pieces:
                    type: integer
                    example: 2
                  description:
                    type: string
                    example: Samsung A51
                  codAmount:
                    type: integer
                    example: 10
                  paymentType:
                    type: string
                    example: COD
                  clientComments:
                    type: string
                    example: Handle With Care
                  value:
                    type: number
                    format: float
                    description: if CC order then required else not
                    example: 100
                  senderInformation:
                    $ref: '#/paths/~1api~1create~1order/post/requestBody/content/application~1json/schema/sender'
                  customer:
                    $ref: '#/paths/~1api~1create~1order/post/requestBody/content/application~1json/schema/admin'
                  sender:
                    type: string
                    example: John
                  weight:
                    type: integer
                    example: 5.2
              sender:
                type: object
                required:
                  - address
                  - city
                  - contactNumber
                properties:
                  address:
                    type: string
                    example: Sulay
                  city:
                    $ref: '#/paths/~1api~1create~1order/post/requestBody/content/application~1json/schema/city1'
                  contactNumber:
                    type: integer
                    example: 966300788877
              city1:
                required:
                  - name
                properties:
                  name:
                    type: string
                    example: Jeddah
              admin:
                type: object
                required:
                  - name
                  - customerAddresses
                  - mobile1
                properties:
                  name:
                    type: string
                    example: David
                  customerAddresses:
                    $ref: '#/paths/~1api~1create~1order/post/requestBody/content/application~1json/schema/Caddress'
                  mobile1:
                    type: integer
                    example: 9612398444
              Caddress:
                required:
                  - address
                  - areaName
                properties:
                  address:
                    type: string
                    example: 358 Riyadh Sulah
                  areaName:
                    type: string
                    example: Riyadh
                  city:
                    $ref: '#/paths/~1api~1create~1order/post/requestBody/content/application~1json/schema/City'
              City:
                required:
                  - name
                properties:
                  name:
                    type: string
                    example: Riyadh
        description: order creation
  /api/fulfilment/order:
    post:
      tags:
        - Fulfilment
      summary: Create Fulfilment Order
      description: Fulfilment Order
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: null
                  httpStatus:
                    type: string
                    example: OK
                  data:
                    $ref: '#/paths/~1api~1fulfilment~1order/post/responses/200/content/application~1json/schema/data'
                data:
                  type: object
                  properties:
                    total:
                      type: integer
                      example: 1
                    totalCreated:
                      type: integer
                      example: 1
                    totalFailed:
                      type: integer
                      example: 0
                    orders:
                      type: array
                      items:
                        $ref: '#/paths/~1api~1fulfilment~1order/post/responses/200/content/application~1json/schema/orderrs'
                orderrs:
                  type: object
                  properties:
                    orderNumber:
                      type: string
                      example: BS-01
                    orderTrackingNumber:
                      type: string
                      example: FUL0000000001
                    status:
                      type: string
                      example: CREATED
                    msg:
                      type: string
                      example: Pending Fulfilment Order created
                    order:
                      type: string
                      example: null
                    type:
                      type: string
                      example: null
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: fail
                  message:
                    type: string
                    example: Duplicate Reference Number For Client
                  httpStatus:
                    type: string
                    example: BAD_REQUEST
                  data:
                    type: string
                    example: null
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - orders
                - products
                - mobile1
                - city
              properties:
                orders:
                  $ref: '#/paths/~1api~1fulfilment~1order/post/requestBody/content/application~1json/schema/orders'
              orders:
                type: array
                items:
                  $ref: '#/paths/~1api~1fulfilment~1order/post/requestBody/content/application~1json/schema/by'
              by:
                type: object
                required:
                  - referenceNumber
                  - description
                  - codAmount
                  - paymentType
                  - clientComments
                  - value
                  - weight
                  - sender
                properties:
                  referenceNumber:
                    type: string
                    example: FUL-01
                  description:
                    type: string
                    example: Samsung Ultra
                  codAmount:
                    type: integer
                    example: 10
                  paymentType:
                    type: string
                    example: CC
                  clientComments:
                    type: string
                    example: Handle With Care
                  value:
                    type: integer
                    example: 0
                  senderInformation:
                    $ref: '#/paths/~1api~1fulfilment~1order/post/requestBody/content/application~1json/schema/sender'
                  weight:
                    type: integer
                    example: 1.3
                  sender:
                    type: string
                    example: John
                  customer:
                    $ref: '#/paths/~1api~1fulfilment~1order/post/requestBody/content/application~1json/schema/admin'
                  products:
                    type: array
                    items:
                      $ref: '#/paths/~1api~1fulfilment~1order/post/requestBody/content/application~1json/schema/products'
              sender:
                type: object
                required:
                  - address
                  - city
                  - contactNumber
                properties:
                  address:
                    type: string
                    example: Jeddah
                  city:
                    $ref: '#/paths/~1api~1fulfilment~1order/post/requestBody/content/application~1json/schema/city1'
                  contactNumber:
                    type: integer
                    example: 966800123654
              city1:
                required:
                  - name
                properties:
                  name:
                    type: string
                    example: Riyadh
              admin:
                type: object
                required:
                  - name
                  - customerAddresses
                  - mobile1
                properties:
                  name:
                    type: string
                    example: David
                  customerAddresses:
                    $ref: '#/paths/~1api~1fulfilment~1order/post/requestBody/content/application~1json/schema/Caddress'
                  mobile1:
                    type: integer
                    example: 966321456987
              Caddress:
                required:
                  - address
                  - areaName
                properties:
                  address:
                    type: string
                    example: 388 Riyadh Sulah
                  areaName:
                    type: string
                    example: Riyadh Town
                  city:
                    $ref: '#/paths/~1api~1fulfilment~1order/post/requestBody/content/application~1json/schema/City'
              City:
                required:
                  - name
                properties:
                  id:
                    type: integer
                    example: 48
                  name:
                    type: string
                    example: Riyadh
              products:
                required:
                  - id
                  - quantity
                properties:
                  id:
                    type: string
                    example: SKU0000000001
                  quantity:
                    type: integer
                    example: 1
        description: Fulfilment Order
  /api/product/:
    post:
      tags:
        - Fulfilment
      summary: Create Fulfilment Product
      description: create Product in Fulfilment
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: null
                  httpStatus:
                    type: string
                    example: OK
                  data:
                    type: array
                    items:
                      $ref: '#/paths/~1api~1product~1/post/responses/200/content/application~1json/schema/data1'
                data1:
                  properties:
                    productNumber:
                      type: string
                      example: BS-01
                    productTrackingSku:
                      type: string
                      example: SKU0000000001
                    status:
                      type: string
                      example: success
                    msg:
                      type: string
                      example: Product Created Successfully
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: null
                  httpStatus:
                    type: string
                    example: OK
                  data:
                    type: array
                    items:
                      $ref: '#/paths/~1api~1product~1/post/responses/400/content/application~1json/schema/data'
                data:
                  properties:
                    productNumber:
                      type: string
                      example: BS-01
                    productTrackingSku:
                      type: string
                      example: null
                    status:
                      type: string
                      example: fail
                    msg:
                      type: string
                      example: Duplicate Reference Number For Client
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - client
                - products
                - data
              properties:
                products:
                  $ref: '#/paths/~1api~1product~1/post/requestBody/content/application~1json/schema/products'
              products:
                type: array
                items:
                  $ref: '#/paths/~1api~1product~1/post/requestBody/content/application~1json/schema/by'
              by:
                type: object
                required:
                  - referenceSKU
                  - brand
                  - weight
                  - height
                  - length
                  - width
                  - description
                  - image
                  - isCompanyProduct
                properties:
                  referenceSKU:
                    type: string
                    example: SK-01
                  brand:
                    type: string
                    example: Samsung
                  weight:
                    type: integer
                    example: 1.3
                  height:
                    type: integer
                    example: 10
                  length:
                    type: integer
                    example: 100
                  width:
                    type: integer
                    example: 5.5
                  description:
                    type: string
                    example: New Edition Available
                  image:
                    type: string
                    example: 'https://storage.googleapis.com/pic.jpg'
                  isCompanyProduct:
                    type: boolean
                    example: true
        description: Create Product
...
  /api/order/schedule:
    post:
      tags:
        - Last Mile
      summary: Order Schedule
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - id
                - customerComments
              properties:
                customer:
                  $ref: '#/paths/~1api~1order~1schedule/post/requestBody/content/application~1json/schema/customer'
                id:
                  type: string
                  example: API0000000001
                customerComments:
                  type: string
                  example: Perfect
              customer:
                type: object
                required:
                  - customerAddresses
                  - mobile1
                properties:
                  customerAddresses:
                    $ref: '#/paths/~1api~1order~1schedule/post/requestBody/content/application~1json/schema/customerAddresses'
                  mobile1:
                    type: integer
                    example: 966321456987
              customerAddresses:
                type: object
                properties:
                  city:
                    $ref: '#/paths/~1api~1order~1schedule/post/requestBody/content/application~1json/schema/city'
                  address:
                    type: string
                    example: Riyadh Town
                  areaName:
                    type: string
                    example: Riyadh
                  latitude:
                    type: integer
                    example: 31.4766839
                  longitude:
                    type: integer
                    example: 74.2766072
                  whatThreeWords:
                    type: string
                    example: riyadh
              city:
                type: object
                properties:
                  name:
                    type: string
                    example: Riyadh
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: null
                  httpStatus:
                    type: string
                    example: OK
                  data:
                    $ref: '#/paths/~1api~1order~1schedule/post/responses/200/content/application~1json/schema/data'
                data:
                  type: object
                  properties:
                    total:
                      type: integer
                      example: 1
                    totalCreated:
                      type: integer
                      example: 1
                    totalFailed:
                      type: integer
                      example: 0
                    orders:
                      type: array
                      items:
                        $ref: '#/paths/~1api~1order~1schedule/post/responses/200/content/application~1json/schema/orders'
                orders:
                  type: object
                  properties:
                    orderNumber:
                      type: string
                      example: API-1
                    orderTrackingNumber:
                      type: string
                      example: API0000000001
                    status:
                      type: string
                      example: success
                    msg:
                      type: string
                      example: order scheduled
                    order:
                      type: string
                      example: null
                    type:
                      type: string
                      example: DROP_SHIP
  /api/get/awb:
    post:
      tags:
        - Last Mile
      summary: AWB
      parameters:
        - in: query
          name: orderid
          required: true
          schema:
            type: string
            example: API0000178748
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                properties:
                  AWB:
                    type: string
                    example: AWB PDF DOWNLOAD
...
  /api/get/all/cities:
    post:
      tags:
        - Cities
      summary: Get Cities
      description: Get Cities
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: null
                  httpStatus:
                    type: string
                    example: OK
                  data:
                    type: array
                    items:
                      $ref: '#/paths/~1api~1get~1all~1cities/post/responses/200/content/application~1json/schema/data'
                data:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 48
                    name:
                      type: string
                      example: Riyadh
                    code:
                      type: string
                      example: RUH
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: fail
                  message:
                    type: string
                    example: Country not found in database
                  httpStatus:
                    type: string
                    example: BAD_REQUEST
                  data:
                    type: string
                    example: null
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - id
              properties:
                id:
                  type: integer
                  description: Country ID
                  example: 39
        description: Get All Cities
...
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic